pyproject.toml
[project]
name = "enroll"
version = "1.2.0"
description = "Monitor Usenet Indexers and notify about open registrations."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-webhook>=1.3.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "loguru-discord>=1.4.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.9.1",
]

[tool.ruff.lint]
select = ["I"]
